Rename the enigmatic message "Unexpected end of expression"
